python/requirements.txt
streamlit==1.32.0
snowflake-connector-python>=3.0.3
requests>=2.31.0
python-dotenv==1.0.1
pyyaml==6.0.1
uuid==1.30 